pyproject.toml
[project]
name = "data-pdf-openai"
version = "0.1.0"
description = ""
authors = [
    {name = "Cezar Maldini",email = "cezar.datascience@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "openai (>=1.61.0,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
